move everything related to player data to its own package
