+++ src/main/java/de/blazemcworld/blazinggames/commands/plural/MemberCommand.java
@@ -26,12 +26,12 @@ import de.blazemcworld.blazinggames.commands.finalizers.ShowNameplatesFinalizer;
 import de.blazemcworld.blazinggames.commands.middleware.EmptyMessageMiddleware;
 import de.blazemcworld.blazinggames.commands.middleware.RequireMemberMiddleware;
 import de.blazemcworld.blazinggames.commands.middleware.RequireSystemMiddleware;
-import de.blazemcworld.blazinggames.utils.DisplayTag;
-import de.blazemcworld.blazinggames.utils.FrontManager;
-import de.blazemcworld.blazinggames.utils.MemberData;
+import de.blazemcworld.blazinggames.players.DisplayTag;
+import de.blazemcworld.blazinggames.players.FrontManager;
+import de.blazemcworld.blazinggames.players.MemberData;
+import de.blazemcworld.blazinggames.players.PlayerConfig;
+import de.blazemcworld.blazinggames.players.PluralConfig;
 import de.blazemcworld.blazinggames.utils.Pair;
-import de.blazemcworld.blazinggames.utils.PlayerConfig;
-import de.blazemcworld.blazinggames.utils.PluralConfig;
 import io.papermc.paper.command.brigadier.CommandSourceStack;
 import io.papermc.paper.command.brigadier.Commands;
 import net.kyori.adventure.text.Component;
