+++ src/main/java/de/blazemcworld/blazinggames/discord/DiscordApp.java
@@ -21,10 +21,10 @@ import club.minnced.discord.webhook.send.WebhookMessageBuilder;
 import de.blazemcworld.blazinggames.BlazingGames;
 import de.blazemcworld.blazinggames.discord.commands.*;
 import de.blazemcworld.blazinggames.discord.eventhandlers.DiscordChatHandler;
-import de.blazemcworld.blazinggames.utils.DisplayTag;
+import de.blazemcworld.blazinggames.players.DisplayTag;
+import de.blazemcworld.blazinggames.players.PlayerConfig;
+import de.blazemcworld.blazinggames.players.PlayerInfo;
 import de.blazemcworld.blazinggames.utils.EmojiRegistry;
-import de.blazemcworld.blazinggames.utils.PlayerConfig;
-import de.blazemcworld.blazinggames.utils.PlayerInfo;
 import net.dv8tion.jda.api.JDA;
 import net.dv8tion.jda.api.JDABuilder;
 import net.dv8tion.jda.api.entities.Member;
