+++ src/main/java/de/blazemcworld/blazinggames/events/handlers/player/PlayerJoinHandler.java
@@ -18,7 +18,7 @@ package de.blazemcworld.blazinggames.events.handlers.player;
 
 import de.blazemcworld.blazinggames.events.base.BlazingEventHandler;
 import de.blazemcworld.blazinggames.items.recipes.RecipeProviders;
-import de.blazemcworld.blazinggames.utils.PlayerConfig;
+import de.blazemcworld.blazinggames.players.PlayerConfig;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.format.TextColor;
 import org.bukkit.event.player.PlayerJoinEvent;
