+++ src/main/java/de/blazemcworld/blazinggames/events/handlers/player/PlayerQuitHandler.java
@@ -17,7 +17,7 @@
 package de.blazemcworld.blazinggames.events.handlers.player;
 
 import de.blazemcworld.blazinggames.events.base.BlazingEventHandler;
-import de.blazemcworld.blazinggames.utils.PlayerConfig;
+import de.blazemcworld.blazinggames.players.PlayerConfig;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.format.TextColor;
 import org.bukkit.event.player.PlayerQuitEvent;
