+++ src/main/java/de/blazemcworld/blazinggames/players/PlayerConfig.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.blazemcworld.blazinggames.utils;
+package de.blazemcworld.blazinggames.players;
 
 import java.util.Properties;
 import java.util.UUID;
